Update project run configurations and upgrade application versions
This commit updates the run configurations for "pim" and "shop", including changes to MavenSettings. The artifactId and version of modules in pom.xml files are also updated. New security property is added to shop's application.properties file. Changes in Docker files reflect the updated versions and the execution of cart service in entrypoint.sh file is slightly adjusted. The overall modifications aim at improving the application setup and running procedures.